RabbitMQ clustering fix for v4.1.2
Addresses issue with Erlang cookie in RabbitMQ clustering by
preseeding the /var/lib/rabbitmq/.erlang.cookie file with the
default value from the cookbooks.

This could be handled in a number of different ways, but this is
the simplest way and takes care of it before the Chef environment
is created.